refactor(runtime): split the runtime engine facade
